Move line break test from lib.rs into line_breaker.rs
To consolidate the tests in line_breaker.rs and keep the lib.rs simple.
